Kernel: Add process groups

+++ qor-os/src/syscalls/kill.rs
@@ -17,14 +17,32 @@ pub fn syscall_kill(proc: &mut super::Process, pid: usize, signal: usize) -> usi
 
     kdebugln!(Syscalls, "PID {} Sending Signal {:?} to PID {}", proc.pid, sig_type, pid);
 
-    if process::scheduler::get_process_manager().as_mut().unwrap().send_signal(
-        pid as u16, 
-        POSIXSignal::new(pid as u16, proc.pid, sig_type)).is_err()
+    if pid != 0
     {
-        errno::ESRCH // Bad pid
+        if process::scheduler::get_process_manager().as_mut().unwrap().send_signal(
+            pid as u16, 
+            POSIXSignal::new(pid as u16, proc.pid, sig_type)).is_err()
+        {
+            errno::ESRCH // Bad pid
+        }
+        else
+        {
+            0
+        }
     }
     else
     {
-        0
+        if process::scheduler::get_process_manager().as_mut().unwrap().send_signal_group(
+            proc.data.process_group_id,
+            proc.pid,
+            POSIXSignal::new(pid as u16, proc.pid, sig_type)).is_err()
+        {
+            errno::ESRCH // Bad pid
+        }
+        else
+        {
+            0
+        }
     }
+    
 }
